+++ source/labyrinths/mechahive.js
@@ -173,7 +173,8 @@ module.exports = new LabyrinthTemplate("Mechahive",
 				"Barrier",
 				"Bow",
 				"Inspiration",
-				"Refreshing Breeze"
+				"Refreshing Breeze",
+				"Universal Solution"
 			],
 			Rare: [
 				"Cleansing Barrier",
@@ -187,7 +188,10 @@ module.exports = new LabyrinthTemplate("Mechahive",
 				"Sweeping Inspiration",
 				"Accelerating Refreshing Breeze",
 				"Supportive Refreshing Breeze",
-				"Swift Refreshing Breeze"
+				"Swift Refreshing Breeze",
+				"Centering Universal Solution",
+				"Evasive Universal Solution",
+				"Harmful Universal Solution"
 			]
 		},
 		Untyped: {